+++ Operation.h
@@ -11,20 +11,20 @@ class Operation
     int operationUserId;
     int operationDate;
     string operationNameFromUser;
-    int operationAmount;
+    float operationAmount;
 
 public:
     void setOperationId(int newOperationId);
     void setOperationuUserId(int newOperationUserId);
     void setOperationDate(int newOperationDate);
     void setOperationNameFromUser(string newOperationNameFromUser);
-    void setOperationAmount(int newOperationAmount);
+    void setOperationAmount(float newOperationAmount);
 
     int getOperationId();
     int getOperationUserId();
     int getOperationDate();
     string getOperationNameFromUser();
-    int getOperationAmount();
+    float getOperationAmount();
 };
 
 #endif
